agent-examples/02-crewai-kickoff-agent/requirements.txt
crewai
python-dotenv
setuptools<81
